requirements.txt
# AI-Powered Blog Generation Service - Production Dependencies
# Pinned to exact versions for reproducibility

# Web Framework
flask==3.0.0
gunicorn==22.0.0

# HTTP and Web Scraping
requests==2.32.3
beautifulsoup4==4.12.3

# AI and ML APIs
openai==1.93.0
anthropic==0.54.0
google-generativeai==0.8.3
google-genai
# Cloud Storage
boto3==1.34.0

# Configuration and Data
python-dotenv==1.0.0
PyYAML==6.0.1

# Optional: AWS Authentication (if needed)
requests-aws4auth==1.3.1 